Bound generic in ParseResult to AsRef<[u8]>
This requires to introduce integer new types eg U8 for u8 which implements
AsRef, and allows to remove the consumed field because it's equal to
`as_ref().len()`.

Reduce stack size of Error to 8 bytes, allowing to speed up computations.

Move ParseResult to own module so that code is obliged to access those via
methods.